ELECTRON-510: Insecure Download Manager Functionality
1. Adds a confirmation dialog before the user can open a file. This confirmation dialog is initiated from the main process.
2. Moves the logic of opening a file from the preload script (no more access
 to a remote site on this logic)
3. Moves the logic of showing a file in the explorer/finder from the preload script (no more access to a remote site on this logic)
4. Creates a few APIs to manage the above
5. Doesn't anymore store the path of a file in the `downloadItems` array in the preload script.
6. Adds a method to add downloads data and clear it from the `downloadItems` array.
7. Some code refactoring.
